Material Builder and Flag Builder reorganization.
