+++ src/main/kotlin/gregtechlite/gtlitecore/api/unification/material/GTLiteFirstDegreeMaterials.kt
@@ -186,6 +186,19 @@ import gregtech.api.unification.material.info.MaterialIconSet.RUBY
 import gregtech.api.unification.material.info.MaterialIconSet.SAND
 import gregtech.api.unification.material.info.MaterialIconSet.SHINY
 import gregtech.api.unification.material.properties.BlastProperty.GasTier
+import gregtechlite.gtlitecore.api.MINUTE
+import gregtechlite.gtlitecore.api.SECOND
+import gregtechlite.gtlitecore.api.TICK
+import gregtechlite.gtlitecore.api.extension.blastProp
+import gregtechlite.gtlitecore.api.extension.cableProp
+import gregtechlite.gtlitecore.api.extension.colorAverage
+import gregtechlite.gtlitecore.api.extension.fluidPipeProp
+import gregtechlite.gtlitecore.api.extension.gas
+import gregtechlite.gtlitecore.api.extension.itemPipeProp
+import gregtechlite.gtlitecore.api.extension.liquid
+import gregtechlite.gtlitecore.api.extension.plasma
+import gregtechlite.gtlitecore.api.extension.rotorProp
+import gregtechlite.gtlitecore.api.extension.toolProp
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ActiniumOxalate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ActiniumSuperhydride
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ActiniumTrihydride
@@ -213,11 +226,14 @@ import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.AmmoniumPertechne
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.AmmoniumSulfate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.AmorphousBoronNitride
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Anorthite
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.AstatineAzide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Augite
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Azurite
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BETSPerrhenate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Baddeleyite
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BariumDichloride
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BariumHydroxide
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BariumManganate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BariumNitrate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BariumOxide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BariumStrontiumTitanate
@@ -235,6 +251,7 @@ import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BlackPhosphorus
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BluePhosphorus
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BlueVitriol
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BoricAcid
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BoronCarbide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BoronTrichloride
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BoronTrifluoride
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BoronTrioxide
@@ -260,6 +277,7 @@ import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.CeriumCarbonate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.CeriumOxide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ChlorinatedSolvents
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ChloroauricAcid
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ChromaticGlass
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ChromiumGermaniumTelluride
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ChromiumGermaniumTellurideMagnetic
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Clinochlore
@@ -292,6 +310,7 @@ import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Fluorite
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Forsterite
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.FranciumBromide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.FranciumCaesiumCadmiumBromide
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.FranciumCarbide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.GSTGlass
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.GadoliniumOxide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.GalliumDioxide
@@ -303,14 +322,18 @@ import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.GermaniumTetrachl
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.GrapheneOxide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HRAMagnesium
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HarmonicPhononMatter
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HassiumTetrachloride
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HeavyAlkaliChloridesSolution
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HeavyConductiveMixture
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HeavyQuarkDegenerateMatter
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HeavyQuarkEnrichedMixture
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HeliumNeon
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Heterodiamond
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HexachloroplatinicAcid
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HexafluorophosphoricAcid
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HexagonalBoronNitride
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HexagonalSiliconNitride
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HolmiumIodide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HolmiumOxide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HydrobromicAcid
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HydrogenPeroxide
@@ -330,6 +353,7 @@ import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.LanthanumEmbedded
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.LanthanumFullereneMixture
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.LanthanumFullereneNanotube
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.LanthanumGalliumManganate
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.LanthanumHexaboride
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.LanthanumOxide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.LanthanumZirconate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.LeadBismuthEutatic
@@ -355,6 +379,7 @@ import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.LuTmYChloridesSol
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.LuTmYVO
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.LutetiumManganeseGermanium
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.LutetiumOxide
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.MOX
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.MagnesiumBromide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.MagnetoResonatic
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ManganeseDifluoride
@@ -398,6 +423,7 @@ import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.PiranhaSolution
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.PlatinumGroupConcentrate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.PlatinumGroupResidue
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Plutonium244
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.PlutoniumDioxide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.PlutoniumPhosphide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.PlutoniumTrihydride
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.PoloniumDioxide
@@ -426,6 +452,8 @@ import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.RadiumDichloride
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.RadiumRadonMixture
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.RadiumSulfate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.RedPhosphorus
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.RheniumHassiumThalliumIsophtaloylbisdiethylthioureaHexafluorophosphate
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.RheniumPentachloride
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.RhodiumTrioxide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.RoastedSphalerite
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.RubidiumChloride
@@ -471,10 +499,12 @@ import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.SodiumPolonate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.SodiumPotassiumEutatic
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.SodiumSeaborgate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.SodiumTellurite
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.SodiumThiocyanate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.SodiumThiosulfate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.SodiumTitanate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.SodiumTrifluoroethanolate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.SodiumVanadate
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.SolarGradeSilicon
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Strontianite
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.StrontiumEuropiumNihonate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.StrontiumFerrite
@@ -490,6 +520,7 @@ import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Tenorite
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.TerbiumOxide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.TetrafluoroboricAcid
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ThalliumBariumCalciumCuprate
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ThalliumChloride
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ThalliumRoentgeniumChloride
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ThalliumSulfate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ThalliumThuliumDopedCaesiumIodide
@@ -524,48 +555,18 @@ import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.YttriumOxide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ZBLANGlass
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ZSM5
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ZincRichSphalerite
+import gregtechlite.gtlitecore.api.unification.material.GTMaterialBuilder.addMaterial
 import gregtechlite.gtlitecore.api.unification.material.info.GTLiteMaterialFlags.DISABLE_CRYSTALLIZATION
 import gregtechlite.gtlitecore.api.unification.material.info.GTLiteMaterialFlags.GENERATE_BOULE
 import gregtechlite.gtlitecore.api.unification.material.info.GTLiteMaterialFlags.GENERATE_NANITE
 import gregtechlite.gtlitecore.api.unification.material.info.GTLiteMaterialFlags.NO_ALLOY_BLAST_RECIPES
 import gregtechlite.gtlitecore.api.unification.material.info.GTLiteMaterialIconSet.BEDROCKIUM
+import gregtechlite.gtlitecore.api.unification.material.info.GTLiteMaterialIconSet.CHROMATIC
 import gregtechlite.gtlitecore.api.unification.material.info.GTLiteMaterialIconSet.GLITCH
 import gregtechlite.gtlitecore.api.unification.material.info.GTLiteMaterialIconSet.MAGNETO
 import gregtechlite.gtlitecore.api.unification.material.info.GTLiteMaterialIconSet.NANOPARTICLES
 import gregtechlite.gtlitecore.api.unification.material.info.GTLiteMaterialIconSet.ROASTED
 import gregtechlite.gtlitecore.api.unification.material.info.GTLiteMaterialIconSet.WHITE_DWARF
-import gregtechlite.gtlitecore.api.MINUTE
-import gregtechlite.gtlitecore.api.SECOND
-import gregtechlite.gtlitecore.api.TICK
-import gregtechlite.gtlitecore.api.extension.blastProp
-import gregtechlite.gtlitecore.api.extension.cableProp
-import gregtechlite.gtlitecore.api.extension.colorAverage
-import gregtechlite.gtlitecore.api.extension.fluidPipeProp
-import gregtechlite.gtlitecore.api.extension.itemPipeProp
-import gregtechlite.gtlitecore.api.extension.rotorProp
-import gregtechlite.gtlitecore.api.extension.toolProp
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.AstatineAzide
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BariumDichloride
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BariumManganate
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BoronCarbide
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ChromaticGlass
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.FranciumCarbide
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HassiumTetrachloride
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HeavyConductiveMixture
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HexafluorophosphoricAcid
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HolmiumIodide
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.LanthanumHexaboride
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.MOX
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.PlutoniumDioxide
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.RheniumHassiumThalliumIsophtaloylbisdiethylthioureaHexafluorophosphate
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.RheniumPentachloride
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.SodiumThiocyanate
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ThalliumChloride
-import gregtechlite.gtlitecore.api.unification.material.info.GTLiteMaterialIconSet.CHROMATIC
-import gregtechlite.gtlitecore.api.extension.liquid
-import gregtechlite.gtlitecore.api.extension.gas
-import gregtechlite.gtlitecore.api.extension.plasma
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.SolarGradeSilicon
 
 object GTLiteFirstDegreeMaterials
 {
