+++ src/main/kotlin/gregtechlite/gtlitecore/api/unification/material/GTLiteSecondDegreeMaterials.kt
@@ -181,6 +181,17 @@ import gregtech.api.unification.material.info.MaterialIconSet.METALLIC
 import gregtech.api.unification.material.info.MaterialIconSet.SAND
 import gregtech.api.unification.material.info.MaterialIconSet.SHINY
 import gregtech.api.unification.material.properties.BlastProperty.GasTier
+import gregtechlite.gtlitecore.api.MINUTE
+import gregtechlite.gtlitecore.api.SECOND
+import gregtechlite.gtlitecore.api.TICK
+import gregtechlite.gtlitecore.api.extension.blastProp
+import gregtechlite.gtlitecore.api.extension.cableProp
+import gregtechlite.gtlitecore.api.extension.fluidPipeProp
+import gregtechlite.gtlitecore.api.extension.gas
+import gregtechlite.gtlitecore.api.extension.itemPipeProp
+import gregtechlite.gtlitecore.api.extension.liquid
+import gregtechlite.gtlitecore.api.extension.rotorProp
+import gregtechlite.gtlitecore.api.extension.toolProp
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Abyssalloy
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ActiniumGroupAlloyA
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ActiniumGroupAlloyB
@@ -193,6 +204,8 @@ import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ArceusAlloy2B
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BETSPerrhenate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BabbitAlloy
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Bedrockium
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BoronFranciumCarbideSuperconductor
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.CaesiumCeriumCobaltIndium
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.CinobiteA243
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.CosmicNeutronium
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.EglinSteel
@@ -209,6 +222,7 @@ import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HastelloyK243
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HastelloyN
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HastelloyX
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HastelloyX78
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HeavyConductiveMixture
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.IncoloyMA813
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.IncoloyMA956
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Inconel625
@@ -227,18 +241,22 @@ import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.MetastableHassium
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.MetastableOganesson
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.MolybdenumDisilicide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Monel500
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.NeutroniumDopedCarbonNanotube
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.NeutroniumSuperconductor
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Nitinol60
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.NobleGasMixture
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.NonMetallicMixtureA
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.NonMetallicMixtureB
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Periodicium
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Pikyonium64B
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.PlutoniumPhosphide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.PreciousMetalAlloy
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.QuantumAlloy
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.RareEarthAlloy
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.RedPhosphorus
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.RefractoryAlloy
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ReneN5
+import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.RheniumHassiumThalliumIsophtaloylbisdiethylthioureaHexafluorophosphate
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.SeaborgiumCarbide
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.SeaborgiumDopedCarbonNanotube
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.SiliconCarbide
@@ -265,27 +283,10 @@ import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.VibraniumTritaniu
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.WatertightSteel
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.Zeron100
 import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.ZirconiumCarbide
+import gregtechlite.gtlitecore.api.unification.material.GTMaterialBuilder.addMaterial
 import gregtechlite.gtlitecore.api.unification.material.info.GTLiteMaterialFlags.NO_ALLOY_BLAST_RECIPES
 import gregtechlite.gtlitecore.api.unification.material.info.GTLiteMaterialIconSet.ENRICHED
 import gregtechlite.gtlitecore.api.unification.material.info.GTLiteMaterialIconSet.HALKONITE
-import gregtechlite.gtlitecore.api.MINUTE
-import gregtechlite.gtlitecore.api.SECOND
-import gregtechlite.gtlitecore.api.TICK
-import gregtechlite.gtlitecore.api.extension.blastProp
-import gregtechlite.gtlitecore.api.extension.cableProp
-import gregtechlite.gtlitecore.api.extension.fluidPipeProp
-import gregtechlite.gtlitecore.api.extension.gas
-import gregtechlite.gtlitecore.api.extension.itemPipeProp
-import gregtechlite.gtlitecore.api.extension.liquid
-import gregtechlite.gtlitecore.api.extension.rotorProp
-import gregtechlite.gtlitecore.api.extension.toolProp
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.BoronFranciumCarbideSuperconductor
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.CaesiumCeriumCobaltIndium
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.HeavyConductiveMixture
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.NeutroniumDopedCarbonNanotube
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.NeutroniumSuperconductor
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.PlutoniumPhosphide
-import gregtechlite.gtlitecore.api.unification.GTLiteMaterials.RheniumHassiumThalliumIsophtaloylbisdiethylthioureaHexafluorophosphate
 import net.minecraft.init.Enchantments
 
 object GTLiteSecondDegreeMaterials
